pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fm-knowledge-service"
version = "1.0.0"
description = "FaultMaven Knowledge Service - Microservice for knowledge base management with RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "chromadb>=0.5.3",
    "sentence-transformers>=3.0.1",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "tenacity>=8.3.0",
    # FaultMaven Core Library
    "fm-core-lib @ git+https://github.com/FaultMaven/fm-core-lib.git@main",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
# Enterprise vector database (for cloud deployments)
pinecone = [
    "pinecone-client>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
